Make independent of OpenHAB

+++ scripts/run.sh
@@ -34,22 +34,24 @@ do
     # commit all files from current dir:
     git add --all .
 
-    THINGS=userdata/jsondb/org.eclipse.smarthome.core.thing.Thing.json
-    SIZE=$(stat -c '%s' ${THINGS} || echo 0)
-    if [ "${SIZE}" -le 1000 ]; then
-        if [ -f ${NOTICED} ]; then
-          echo "Resetting ${THINGS} and removing ${NOTICED}"
-          git reset
-          git checkout HEAD ${THINGS}
-          rm -f ${NOTICED}
+    if [ -d userdata/jsondb ]; then
+        THINGS=userdata/jsondb/org.eclipse.smarthome.core.thing.Thing.json
+        SIZE=$(stat -c '%s' ${THINGS} || echo 0)
+        if [ "${SIZE}" -le 1000 ]; then
+            if [ -f ${NOTICED} ]; then
+                echo "Resetting ${THINGS} and removing ${NOTICED}"
+                git reset
+                git checkout HEAD ${THINGS}
+                rm -f ${NOTICED}
+            else
+                echo "Waiting for OpenHAB to notice empty things"
+            fi
+            sleep 10
+            continue
         else
-          echo "Waiting for OpenHAB to notice empty things"
+            echo "Removing orphan ${NOTICED}"
+            rm -f ${NOTICED}
         fi
-        sleep 10
-        continue
-    else
-        echo "Removing orphan ${NOTICED}"
-        rm -f ${NOTICED}
     fi
 
     # commit with custom message:
